Revert "feat: add image diffing & testing script (#2)"
This reverts commit e4efc1a6706bb6bd552d163fcebb32665b8569fb.
