+++ .env.TEMPLATE
@@ -1,4 +1,3 @@
-# all values below are required
 code_webhook="https://discord.com/api/webhooks/.../......"
 color_webhook="https://discord.com/api/webhooks/.../......"
 icons_webhook="https://discord.com/api/webhooks/.../......"
