+++ .gitignore
@@ -176,5 +176,4 @@ dist
 
 # Custom
 tmp
-*.env
-src/canvas/mock/*
+.env
