--- local.env.TEMPLATE
@@ -1,6 +1,0 @@
-# used in "bun run update:test"
-# all values below are optional, they are overrides for the base .env file
-code_webhook="https://discord.com/api/webhooks/.../......"
-color_webhook="https://discord.com/api/webhooks/.../......"
-icons_webhook="https://discord.com/api/webhooks/.../......"
-NODE_ENV="test"
