+++ src/update/tasks/diffs.ts
@@ -2,7 +2,28 @@ import { prevFiles } from "..";
 import { handleShellErr, type Progress } from "../util";
 import { join } from "path";
 import Color from "color";
-import { DiffEnum, type Diff } from "../../types";
+
+export type OutDiffs = NonNullable<Awaited<ReturnType<typeof diffs>>>;
+
+export enum DiffEnum {
+  Added,
+  Changed,
+  Removed,
+}
+
+export type Diff =
+  | {
+      change: DiffEnum.Added;
+      cur: string;
+    }
+  | {
+      change: DiffEnum.Changed;
+      old: string;
+      cur: string;
+    }
+  | {
+      change: DiffEnum.Removed;
+    };
 
 type RawColors = Record<string, string>;
 const diffRaw = async (progress: Progress) => {
@@ -30,8 +51,7 @@ const diffRaw = async (progress: Progress) => {
         cur: newRaw[raw],
       });
   for (const raw of Object.keys(oldRaw))
-    if (!newRaw[raw])
-      changes.set(raw, { change: DiffEnum.Removed, old: oldRaw[raw] });
+    if (!newRaw[raw]) changes.set(raw, { change: DiffEnum.Removed });
 
   progress.update("diff_raw", true);
   return changes;
@@ -67,7 +87,7 @@ const diffSemantic = async (progress: Progress) => {
         `${key}.${k}`,
         added
           ? { change: DiffEnum.Added, cur: colorify(v) }
-          : { change: DiffEnum.Removed, old: colorify(v) }
+          : { change: DiffEnum.Removed }
       )
     );
 
@@ -97,7 +117,6 @@ const diffSemantic = async (progress: Progress) => {
         if (!newSemantic[sem][clir])
           changes.set(`${sem}.${clir}`, {
             change: DiffEnum.Removed,
-            old: colorify(oldSemantic[sem][clir]),
           });
 
   progress.update("diff_semantic", true);
@@ -130,8 +149,7 @@ const diffIcons = async (progress: Progress) => {
         cur: newIcons[icon].hash,
       });
   for (const icon of Object.keys(oldIcons))
-    if (!newIcons[icon])
-      changes.set(icon, { change: DiffEnum.Removed, old: oldIcons[icon].hash });
+    if (!newIcons[icon]) changes.set(icon, { change: DiffEnum.Removed });
 
   progress.update("diff_icons", true);
   return changes;
